Refactored some class names and move them to another package
