+++ src/main/java/se/myhappyplants/server/controller/StartServer.java
@@ -3,16 +3,20 @@ package se.myhappyplants.server.controller;
 import se.myhappyplants.server.model.LightCalculator;
 import se.myhappyplants.server.model.WaterCalculator;
 import se.myhappyplants.server.services.PlantRepository;
+import se.myhappyplants.server.services.ServerConnection;
 import se.myhappyplants.server.services.UserPlantRepository;
 import se.myhappyplants.server.services.UserRepository;
 
 /**
  * Created by: Frida Jacobson, Eric Simonson, Anton Holm, Linn Borgström, Christopher O'Driscoll
+ * Updated by: Linn Borgström, 2021-05-13
  */
 public class StartServer {
-    public static void main(String[] args) throws Exception {
-        Controller controller = new Controller();
+    public static void main(String[] args) {
+        UserRepository userRepository = new UserRepository();
         PlantRepository plantRepository = new PlantRepository(new LightCalculator(), new WaterCalculator());
-        new Server(2555, new UserRepository(), plantRepository, new UserPlantRepository(plantRepository));
+        UserPlantRepository userPlantRepository = new UserPlantRepository(plantRepository);
+        ServerController serverController = new ServerController(userRepository,userPlantRepository,plantRepository);
+        new ServerConnection(2555,serverController);
     }
 }
